scripts/requirements.txt
# Benchmark dependencies
openai>=1.0
anthropic>=0.18
google-genai>=1.0

# For seed script
supabase>=2.3.0
python-dotenv>=1.0.0

